move some logical expression type-coercion logic from physical-expr to expr
