+++ datafusion/physical-expr/src/expressions/mod.rs
@@ -60,26 +60,19 @@ pub mod helpers {
 
 pub use approx_distinct::ApproxDistinct;
 pub use approx_median::ApproxMedian;
-pub use approx_percentile_cont::{
-    is_approx_percentile_cont_supported_arg_type, ApproxPercentileCont,
-};
+pub use approx_percentile_cont::ApproxPercentileCont;
 pub use approx_percentile_cont_with_weight::ApproxPercentileContWithWeight;
 pub use array_agg::ArrayAgg;
-pub use average::is_avg_support_arg_type;
-pub use average::{avg_return_type, Avg, AvgAccumulator};
+pub use average::{Avg, AvgAccumulator};
 pub use binary::{binary, binary_operator_data_type, BinaryExpr};
 pub use case::{case, CaseExpr};
 pub use cast::{
     cast, cast_column, cast_with_options, CastExpr, DEFAULT_DATAFUSION_CAST_OPTIONS,
 };
 pub use column::{col, Column};
-pub use correlation::{
-    correlation_return_type, is_correlation_support_arg_type, Correlation,
-};
+pub use correlation::Correlation;
 pub use count::Count;
-pub use covariance::{
-    covariance_return_type, is_covariance_support_arg_type, Covariance, CovariancePop,
-};
+pub use covariance::{Covariance, CovariancePop};
 pub use cume_dist::cume_dist;
 
 pub use distinct_expressions::{DistinctArrayAgg, DistinctCount};
@@ -98,13 +91,10 @@ pub use nullif::{nullif_func, SUPPORTED_NULLIF_TYPES};
 pub use rank::{dense_rank, percent_rank, rank};
 pub use row_number::RowNumber;
 pub use stats::StatsType;
-pub use stddev::{is_stddev_support_arg_type, stddev_return_type, Stddev, StddevPop};
-pub use sum::is_sum_support_arg_type;
-pub use sum::{sum_return_type, Sum};
+pub use stddev::{Stddev, StddevPop};
+pub use sum::Sum;
 pub use try_cast::{try_cast, TryCastExpr};
-pub use variance::{
-    is_variance_support_arg_type, variance_return_type, Variance, VariancePop,
-};
+pub use variance::{Variance, VariancePop};
 
 /// returns the name of the state
 pub fn format_state_name(name: &str, state_name: &str) -> String {
