+++ datafusion/physical-expr/src/expressions/stddev.rs
@@ -42,42 +42,6 @@ pub struct StddevPop {
     expr: Arc<dyn PhysicalExpr>,
 }
 
-/// function return type of standard deviation
-pub fn stddev_return_type(arg_type: &DataType) -> Result<DataType> {
-    match arg_type {
-        DataType::Int8
-        | DataType::Int16
-        | DataType::Int32
-        | DataType::Int64
-        | DataType::UInt8
-        | DataType::UInt16
-        | DataType::UInt32
-        | DataType::UInt64
-        | DataType::Float32
-        | DataType::Float64 => Ok(DataType::Float64),
-        other => Err(DataFusionError::Plan(format!(
-            "STDDEV does not support {:?}",
-            other
-        ))),
-    }
-}
-
-pub fn is_stddev_support_arg_type(arg_type: &DataType) -> bool {
-    matches!(
-        arg_type,
-        DataType::UInt8
-            | DataType::UInt16
-            | DataType::UInt32
-            | DataType::UInt64
-            | DataType::Int8
-            | DataType::Int16
-            | DataType::Int32
-            | DataType::Int64
-            | DataType::Float32
-            | DataType::Float64
-    )
-}
-
 impl Stddev {
     /// Create a new STDDEV aggregate function
     pub fn new(
@@ -342,17 +306,6 @@ mod tests {
             ScalarValue::from(std::f64::consts::SQRT_2),
             DataType::Float64
         )
-    }
-
-    #[test]
-    fn test_stddev_return_data_type() -> Result<()> {
-        let data_type = DataType::Float64;
-        let result_type = stddev_return_type(&data_type)?;
-        assert_eq!(DataType::Float64, result_type);
-
-        let data_type = DataType::Decimal(36, 10);
-        assert!(stddev_return_type(&data_type).is_err());
-        Ok(())
     }
 
     #[test]
