+++ datafusion/physical-expr/src/expressions/variance.rs
@@ -48,42 +48,6 @@ pub struct VariancePop {
     expr: Arc<dyn PhysicalExpr>,
 }
 
-/// function return type of variance
-pub fn variance_return_type(arg_type: &DataType) -> Result<DataType> {
-    match arg_type {
-        DataType::Int8
-        | DataType::Int16
-        | DataType::Int32
-        | DataType::Int64
-        | DataType::UInt8
-        | DataType::UInt16
-        | DataType::UInt32
-        | DataType::UInt64
-        | DataType::Float32
-        | DataType::Float64 => Ok(DataType::Float64),
-        other => Err(DataFusionError::Plan(format!(
-            "VAR does not support {:?}",
-            other
-        ))),
-    }
-}
-
-pub fn is_variance_support_arg_type(arg_type: &DataType) -> bool {
-    matches!(
-        arg_type,
-        DataType::UInt8
-            | DataType::UInt16
-            | DataType::UInt32
-            | DataType::UInt64
-            | DataType::Int8
-            | DataType::Int16
-            | DataType::Int32
-            | DataType::Int64
-            | DataType::Float32
-            | DataType::Float64
-    )
-}
-
 impl Variance {
     /// Create a new VARIANCE aggregate function
     pub fn new(
@@ -424,17 +388,6 @@ mod tests {
             ScalarValue::from(2_f64),
             DataType::Float64
         )
-    }
-
-    #[test]
-    fn test_variance_return_data_type() -> Result<()> {
-        let data_type = DataType::Float64;
-        let result_type = variance_return_type(&data_type)?;
-        assert_eq!(DataType::Float64, result_type);
-
-        let data_type = DataType::Decimal(36, 10);
-        assert!(variance_return_type(&data_type).is_err());
-        Ok(())
     }
 
     #[test]
